Add some inertia and additional randomness to the RationalConsumer
Allow NSPs with surpluses a specified (nspCandidateFracOfBestThreshold)
fraction of the top edge network candidate to be considered.  The
weights of all candidates are based on how far away they are from the
best NSP candidate within the threshold range.  The incumbent provider
is provided an additive advantage (nspIncumbentAdditiveAdvantage).

The simternet class and property file were updated to add the two
afforementioned new parameters. 

The WeightedIndex class was added to handle the weighting and selection.